backend/requirements/base.txt
Django==3.2.9
djangorestframework==3.12.4
django-filter==21.1
django-cors-headers==3.10.0
psycopg2-binary==2.9.1
neo4j==4.4.0
neo4j-driver==4.3.6
networkx==2.6.3
flake8==4.0.1
requests==2.26.0
beautifulsoup4==4.10.0
pyppeteer==0.2.6
leidenalg==0.8.8
python-igraph==0.9.8
scikit-learn==1.0.1
numpy==1.21.4
